requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types ==0.6.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.4
frozenlist==1.4.1
greenlet==3.0.3
ibm-cos-sdk ==2.13.4
ibm-cos-sdk-core==2.13.4
ibm-cos-sdk-s3transfer==2.13.4
ibm-watson-machine-learning==1.0.353
ibm-watsonx-ai==0.2.3
idna ==3.6
importlib_metadata==7.1.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.14
langchain-community==0.0.31
langchain-core==0.1.38
langchain-ibm==0.1.3
langchain-text-splitters==0.0.1
langsmith==0.1.38
lomond==0.3.3
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions ==1.0.0
numpy==1.26.4
orjson==3.10.0
packaging==23.2
pandas==1.5.3
pip==24.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil ==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests ==2.31.0
setuptools==65.5.0
six==1.16.0
SQLAlchemy==2.0.29
tabulate==0.9.0
tenacity==8.2.3
typing_extensions==4.10.0
typing-inspect==0.9.0
urllib3==2.1.0
yarl ==1.9.4
zipp ==3.18.1